requirements/wip.txt
alembic==0.8.8
beautifulsoup4==4.5.1
coverage==4.2
decorator==4.0.10
factory-boy==2.7.0
fake-factory==0.7.2
flake8==3.0.4
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
mccabe==0.5.2
nose==1.3.7
PasteDeploy==1.5.2
-e git+git@github.com:cdmb/pergunta-camara.git@407b76567c7ce86feb36e136745d1f993c875c71#egg=pergunta_camara
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
py==1.4.31
pycodestyle==2.0.0
pydocstyle==1.1.1
pyflakes==1.3.0
Pygments==2.1.3
pylama==7.2.3
pyramid==1.7.3
pyramid-debugtoolbar==3.0.5
pyramid-ipython==0.2
pyramid-jinja2==2.6.2
pyramid-mailer==0.14.1
pyramid-mako==1.0.2
pyramid-tm==1.0.1
pytest==3.0.3
pytest-cov==2.4.0
python-dateutil==2.6.0
python-editor==1.0.1
repoze.lru==0.6
repoze.sendmail==4.2
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.3
SQLAlchemy-Utils==0.32.9
traitlets==4.3.1
transaction==1.7.0
translationstring==1.3
venusian==1.0
waitress==1.0.1
wcwidth==0.1.7
WebOb==1.6.2
WebTest==2.0.23
zope.deprecation==4.2.0
zope.interface==4.3.2
zope.sqlalchemy==0.7.7
